Apply feature/nine-one onto main (#497)
* Executable does not leave terminal state (#494)

* Miscellaneous fixes (#495)

* Better check for Executable that was never started

* Do not wait on child process i/o

For processes that we launch, we only capture root process i/o. There is no point to wait for child processes i/o to be completed.

(this is slightly reworked change that David originally did as part of https://github.com/microsoft/usvc-apiserver/pull/486)

* Do not send a (late) run session change notification if the session is finished

* Remove unnecessary generic type constraint

(Go compiler got smarter)

* Simplify stopSingleProcess() epilogue

(thanks for the suggestion, Devin)

* Misc fixes part 2 (#496)

* Do not debounce IDE run session notifications

It is not really necessary--there will not be so many notifications to warrant the debouncer use.
But we do want to retain the asynchronous nature of NotifyRunChangedAsync() and NotifyRunCompletedAsync(), so we will use a work queue with concurrency == 1 instead.

* Fix linter issues

* Update linter version

* Retry removal of the log files

The API server may be holding onto the files as it is shutting down the log streams, so it is worth retrying the operation. Lifetime context is used to make sure the retries do not impede DCP shutdown.

* Close resources associated with (closed) log streams more eagerly

This is mostly a re-application of https://github.com/microsoft/usvc-apiserver/pull/488, but with some extra change to avoid a race condition when handling follow writer errors.

* Use  new process group when forking DCP/DCPCTRL

Re-applies https://github.com/microsoft/usvc-apiserver/pull/492. Contrary to what that PR says, this does NOT fully insulate us from being a "child process" of Aspire app host, but being in a separate process group is what we want.

* Refactor container runtime status check for better readability

* Re-introduce the "runtime unhealthy" container state
(re-applies https://github.com/microsoft/usvc-apiserver/pull/490)

* Fix deadlock in subscription package
The deadlock scenario is as follows
Goroutine 1: Calls SubscriptionSet.Notify(), which takes a lock on SubscriptionSet and then tries to call Notify() on each Subscription belonging to the set (which internally takes the Subscription lock).
Goroutine 2: Calls Subscription.Cancel() which takes the Subscription lock, and then calls SubscriptionSet.onSubscriptionCancelled(), which tries to take the SubscriptionSet lock.
The fix prevents SubscriptionSet from calling Subscription.Notify() while holding the SubscriptionSet lock AND prevents Subscription from calling SubscriptionSet.onSubscriptionCancelled() while holding the Subscription lock.
This was discovered by health subsystem tests, but the bug also affects Docker/Podman interaction.

* Simplify follow writer
The atomic "done" member and the finish() method are simply unnecessary

* Fix typo

* Build release with runtime protections enabled

* Eliminate race condition in tests

The problem with the code before the fix is that the TestIdeRunner is holding onto the run/RunInfo object and is modifying it during subsequent method invocations. This can get into a race with the OnStartupCompleted event handler. Creating a deep copy for the event handler prevents this race.